Add more unit test to roles package
